Add: provide `getCollisionLocationsInArea()` function for map (#7707)
#### Brief overview of PR changes/additions
Adds a function that, given an area ID returns a table containing
sub-tables which contain the coordinates of any location in that area
where there is more than one room. Feeding one of those triplets into
`getRoomsByPosition(...)` will then identify the rooms that are clashing
(and which are now marked up in the 2D mapper).

#### Motivation for adding to Mudlet
To aide those creating maps so they can avoid creating maps where rooms
are co-located, without them having to iterate through the rooms in an
area themselves. (I feel it would be more efficient to gather the
required information in the C++ core in one swell foop rather than
repeatedly interrogating the C++ core for the position of each room in
an area).

#### Other info (issues closed, discussion etc)
None

---------

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>