Fix: slice up QKeySequences correctly for Mute/Unmute (#7610)
#### Brief overview of PR changes/additions
Adds a character to split up the OS dependent string `"Alt+K"` at the
`+`.

#### Motivation for adding to Mudlet
PR #7440 tried to improve the UI by showing a message about which keys
to press to undo a Mute/Unmute action the first time one of those is
performed. Unfortunately the choice of the empty string as the separator
in `QString::split(...)` means that the text is split up on a per
character basis - and then reassembled using just double-quoting
`QKeySequence::toString(QKeySequence::NativeText)` so that the above
example was shown as `<A>+<l>+<t>+<+>+<K>` instead of
`"Alt K"` as it will now be...

#### Other info (issues closed, discussion etc)
This was reported by **@bcross32** in the Mudlet Discord
**#Accessibility** channel on 2024/12/27T23:18:

https://www.discord.com/channels/283581582550237184/677242749514416170/1322342815866294373

---------

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>