Infrastructure: sort QGridLayout/polish items in `profilePreferences.ui` file (#7787)
#### Brief overview of PR changes/additions
* Reorder back into ascending row/columns the items in the QGridLayout.
* Also adds better names to some of them which default (back) to
`gridLayout_#` if freshly created and which can then clash if separate
PRs are merged with identical such names.
* Also unify the presentation of tool-tips in this form (so they are all
formatted as wrappable text sentences).
* Add "buddy" connections between labels and the buttons or other
interactive widget they are associated with for labels that have not so
far been so linked but which can be.
* Reorder widgets "tab-order" to accommodate recent changes.

#### Motivation for adding to Mudlet
It is almost impossible to track changes via `git`/`diff` in `.ui` forms
if items in a `QGridLayout` are edited via the `Qt Designer` tool as it
does not preserve the order. We did try and automate this in our CI but
it was not reliable IIRC - so it falls back to manual edits if PR
creators don't do it at the time of their PRs. Other polishes done for
general improvements.

#### Other info (issues closed, discussion etc)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>