Improve: make Notepad font track Main Console one (#7800)
#### Brief overview of PR changes/additions
The font (including size) used for the main console font is now applied
to the Notepad window.

#### Motivation for adding to Mudlet
Raised by the Discord user `@missionz3r0` in our Discord `#help` channel
in the message:

https://wwww.discord.com/channels/283581582550237184/283582068334526464/1357767687471632466
> **missionz3r0** — 2025/04/04 at 17:24 UTC
> 
> Is there a way to change the font  settings in the Notepad?
> 
> Neither the editor or main settings affect it.
> 
> **SlySven** ☕ — 2025/04/04 at 19:07 UTC
> The "Main" setting might affect it if you restart Mudlet after
changing it... but I'll look to see if it can be made to track the main
console font in the future.

#### Other info (issues closed, discussion etc)
I also discovered that there was a dead (unused) method:
`(std::pair<bool, QString>) Host::setDisplayFont(const QString&)` which
I could remove and that I could restructure:`(void)
Host::setDisplayFontFromString(const QString&)` so that a redundent
duplicate call to `(void) Host::updateConsolesFont()` could be removed.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>