Fix: restore use of F3 for keybindings when not used for searching (#7783)
#### Brief overview of PR changes/additions
Only create (and destroy if not wanted) shortcuts used to provide F3 and
Shift+F3 backwards and forwards search in the main console for those
with accessibility requirements.

#### Motivation for adding to Mudlet
I was using the F3 key for a keybinding and after #7579 I found that it
stopped working - even when I disabled the functionality that it
offered. It turns out that the enabling/disabling steps that the creator
used were not sufficient as it left the "shortcuts" in place even though
they didn't do anything and that was enough to bypass the normal
handling of that key.

#### Other info (issues closed, discussion etc
There was a superfluous initialisation of the `(bool)
Host::mF3SearchEnabled` flag in the Host initialisation list which was
also out of order and producing a compilation warning.

Also cleans up some code miss-formatting (omission of braces around `if`
code blocks and no new line after the condition part).

---------

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>