Infrastructure: update Linux and MacOS builds to Qt 6.9.0 (#7808)
#### Brief overview of PR changes/additions
Update the Linux and MacOS CI/CB systems to build with Qt 6.9.0 which is
the latest released version.

#### Motivation for adding to Mudlet
I've recently updated the codebase to compile with this version as the
Windows CI/CB has automagically updated to that version and was breaking
until I fixed it in #7805 so it makes sense to change those other OSes
to match.

#### Other info (issues closed, discussion etc)

---------

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>