Infrastructure: update Linux runner to Ubuntu 22.04 (#7813)
#### Brief overview of PR changes/additions
Change the Linux Ci/CB runner used to make the AppImage for that OS from
the Ubuntu 20.04 LTS one to the 22.04 one.

#### Motivation for adding to Mudlet
This is unavoidable because the older one has been removed.

#### Other info (issues closed, discussion etc)
There was some intransigence by the maintainer of the `linuxdeployqt`
tool we use to make the AppImage it was not being updated to work with
the 22.04 one until the 20.04 reached its official EOL next month.

That has now changed but it is not clear yet that the AppImages produced
in the newer runner will work on systems with the older `libc` that was used
in the 20.04 version and others such as Debian 11 ("Bullseye") and
Devuan 4 ("Chimaera").

---------

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>