Infrastructure: clean up missing line-feeds at start of methods (#7803)
#### Brief overview of PR changes/additions
This PR adds a missing Line-Feed to those methods that are missing them
after the closing `)` or the `) const` of the arguments to that method.

#### Motivation for adding to Mudlet
We intend for the opening `{` to be in the first column of the next line
but some PRs in the past have added methods (functions) where this
hasn't been done and this is not something that the `clang-format` that
we bundle can fix-up for us automatically.

#### Other info (issues closed, discussion etc)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>