Infrastructure: remove redundant ';'s from Q_UNUSED(...) macro (#7825)
#### Brief overview of PR changes/additions
In all places in our code where we use `Q_UNUSED(...)` to silence a
warning about an argument to a method/function that isn't used this PR
removes any following `;` as this macro does NOT need it.

#### Motivation for adding to Mudlet
For consistency across the entirety of our code, so that all our usage
of this macro is "correct" and doesn't include something that is
unneeded.

#### Other info (issues closed, discussion etc)
There are places in third-party code that does include it but it isn't
our job to clean those up!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>