Fix: vertical alignment of glyphs (#7840)
#### Brief overview of PR changes/additions
Revises the code I originated in #7766 to also centre the text
vertically within the space allocated for it as well as horizontally.

#### Motivation for adding to Mudlet
Large emojis were being drawn too far down and then being clipped by the
following line (on MacOS) - this made a regression from how things were
before 7766.

#### Other info (issues closed, discussion etc)
This will introduce the fix I posited in and also close #7839.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>