fix: Lua local variable declaration leaks into global (#7853)
It'll be easier to review this commit by commit instead of as a whole.

It'd also be preferable to squash merge. My commit naming conventions
changed as I went.


### Brief overview of PR changes/additions
A number of functions, methods, and one file on the lua side of things
leaked local variable declarations into the global namespace. This cures
the current leaks.

### Motivation for adding to Mudlet
Leaking local variables into global space can have unforeseen side
affects. If two functions leak the same variable name, calling one might
affect a call to the other. `Geyser.remove` specifically made me wince.

I did not see any cases where this was actively an issue, but thought it
best to remove any chance of it happening.

### Other info (issues closed, discussion etc)

#### Thoughts on future improvements
I'd like to go through to remove *all* global declarations, including at
the file level. Instead, exposing anything global in `LuaGlobal`. I'll
put the gist of why here, but I can make a proper issue if there is
support for it.

Because:
* It'd make any further mistakes like this far more visible
* diagnostics tools will only point at actual issues rather than every
global object.
* Example: here is VS Code complaining about `db` being global, when we
obviously want it to be global. There are thousands of warnings like
this hiding the actual leaks.

![image](https://github.com/user-attachments/assets/5d72d8d6-e768-4314-b219-dc41e1102efd)

* and increase code readability.
* Mudlet devs would need to require modules on this side of things in
order to use them.
* note: this will only affect how Mudlet devs use the functions. Because
`LuaGlobal` will be loaded in for users(which dumps module contents into
the global space), this change should only affect our side of things.
    * Example: calls to `replace` would become `GUIUtils.replace`

---------

Co-authored-by: Zooka <136661366+ZookaOnGit@users.noreply.github.com>