fix: luaGlobalpath and revise LuaGlobal.lua file (#7865)
### Brief overview of PR changes/additions
* fix: luaGlobalPath's fallback path to exclude `LuaGlobal.lua`
* revise: LuaGlobal.lua

### Motivation for adding to Mudlet
* if we end up using the fallback path, the path we'll be trying to
import a package from would be
`[some_dir]/LuaGlobal.lua/[package_file_name].lua`.
* Reasoning through the logic of LuaGlobal was a tad difficult. There
were many branches in the logic and it was split up throughout the file.
With the revisions, it's easier to reason through the logic (at least I
think so). It's what lead me to notice the bug with the fallback path.

### Other info (issues closed, discussion etc)
This is a cherry-picked out of an experiment I was working on to reduce
the use of globals in the lua files. The idea was to expose them in
LuaGlobal but have everything else as modules. But, when I started
digging into LuaGlobal, I ended up making the changes you see here
before getting on with what I was trying to do.