improve: unique constraints for lua db (#7852)
### Brief overview of PR changes/additions
1. Added validation logic for `_unique` and `validations`
2. Removed explicit creation of unique indexes on a sql table
   * replaced with unique constraints
3. Improved unique constraint creation during table creation

### Motivation for adding to Mudlet
1. Running setup only for a database's creation to fail can and will
leave the state of `db` in a dirty state. It's better to error early and
with detail than it is for a dev to iterate over each error as they
stumble into it.
2. Prior logic had the database creating column constraints and unique
indexes. The difference between them is negligible. They both will get
the job done. The problem is that the database will create both, ending
in a doubling of work when verifying uniqueness.
3. So, I opted to remove unique indexes in favor unique constraints
(column and table).

### Other info (issues closed, discussion etc)

#### Issues Closed
https://github.com/Mudlet/Mudlet/issues/7085

#### Downside
Unlike a unique index, a unique constraint is not searchable. However, I
think that's a side affect rather than intended behavior. If I or
someone else wants to search through an index, we'd reach for `_index`.

---------

Co-authored-by: Zooka <136661366+ZookaOnGit@users.noreply.github.com>