Fix: db:Timestamp's usage throughout the db module (#7871)
### Brief overview of PR changes/additions
* Add: tests cases for `db:Timestamp`
* Fix: `datetime:calculate_UTCdiff` should NOT include DST into its
offset.
* Fix: `datetime:parse` to use the source's time instead of current time
to determine DST.
* Fix: `db:Timestamp("CURRENT_TIMESTAMP")` so it generates the correct
SQL during table inserts.
* Fix: `db:_coerce_sheet`'s conversion of Null timestamps so they are
proper `db:Timestamp` tables.
* Fix: `db.__SheetMT`'s mistreatement of `db:Timestamp(nil)` as a
non-timestamp field.

### Motivation for adding to Mudlet
When we put a value into the database we should get the same value out.
As the logic is right now, if we looped back and forth between fetching
and updating a timestamp value from sqlite, it'll change on every
iteration of the loop by the locale's offset from UTC, and their
daylight savings value if the timestamp is far enough in the past.

### Other info (issues closed, discussion etc)
#### Discord Discussion:

https://discord.com/channels/283581582550237184/283582439002210305/1376590435597947020

##### missionz3r0 `May 26, 2025 @ 1158 EST`
> Well, I just uncovered what was going on here.
> 
> When I put a time *into* the database, it's treated as a UTC time.
Sqlite will also give it back in UTC. But then `db._coerce_sheet` turns
it into a localtime. (despite what the comments say)
> ```lua
>         elseif field.type == "datetime" then
>           -- the value, tbl[k], is currently in a UTC timestamp
>           local localtime = datetime:parse(tbl[k], nil, true)
> -- convert it into a UTC timestamp as datetime:parse parses it in the
local time context
> tbl[k] = db:Timestamp(localtime +
datetime:calculate_UTCdiff(localtime))
>         end
> ```
> 
> I'm not seeing anything in `datetime:parse` that looks like it's
converting to a local time. So what it is really doing is taking a UTC
timestamp and turning it into local time.
> 
> Which does a funny thing. Everytime you fetch a row and use it for an
update, the timestamp will change by your UTC offset.

---------

Co-authored-by: Zooka <136661366+ZookaOnGit@users.noreply.github.com>