Infrastructure: adjust name of TConsole::setFont(...) (#7897)
#### Brief overview of PR changes/additions
Renames `(bool) TConsole::setFont(const QString&)` to `setFontName`
because it does actually do the latter (stores the NAME of the font to
use) and could otherwise be mistaken for the base class `(void)
QWidget::setFont(const QFont&)` method.


#### Motivation for adding to Mudlet
To reduce the chance of misinterpreting what the method does in
comparison with what the base class function performs.

#### Other info (issues closed, discussion etc)
It is not impossible that we might restructure things so that we send a
reference to an actual font instance rather than the name of a font in
this class. Such a `setFont` method then would then have the `override`
"label" applied to it so as to flag that it is replacing the base class
function, which the method renamed in this PR does **not**.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>