Fix: get Ctrl+# to work on AZERTY keyboards (#7915)
#### Brief overview of PR changes/additions
Revise processing of keyboard events so that users of French AZERTY
keyboards can still pass "number" keys into Mudlet. That type of
keyboard layout requires the use of the SHIFT modifier to enter the
numeric value on the top row (excluding functions keys) of the keyboard.

#### Brief overview of PR changes/additions
Allow French users to switch between multiple profiles when
multi-playing.

#### Other info (issues closed, discussion etc)
This can be tested from "US" and other "English" keyboards by selecting
a "FR (AZERTY)" layout *without* having to source a specimen of that
particular keyboard (you do have to remember to hold the SHIFT key down
as well as the CTRL one when pressing a number key)!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>