improve: convert HTML currently in script editor error messages sent to screen readers (#7507)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Converts HTML to plain text in strings that are sent to screen readers.
Things like the error message in the script editor have things like bold
in them, and any user defined item name with a less than sign in it such
as a trigger named `hp < 200` should be escaped properly at this point
thanks to earlier updates that were focused on proper displays.

#### Motivation for adding to Mudlet
This was a goal after #6795

#### Other info (issues closed, discussion etc)

---------

Co-authored-by: Zooka <136661366+ZookaOnGit@users.noreply.github.com>