Fix: Suppress a blank line after triggered echo with newline suffix (#7923)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Added a check for whether new lines were added from trigger
processing/wrapping and if the last line in the buffer is blank. If new
lines were added and the last line in the buffer is empty,
TBuffer::translateToPlainText will skip appending an additional blank
line.

#### Motivation for adding to Mudlet
A discord user mentioned the wrap-overhaul causing their echoes to
behave differently.

#### Other info (issues closed, discussion etc)
4.19.1:

![image](https://github.com/user-attachments/assets/51be1432-b534-4f10-991c-1f57a89714a7)

After PR #7714:

![image](https://github.com/user-attachments/assets/27a74309-8d48-4933-8194-788b7ced77ac)

After this PR:

![image](https://github.com/user-attachments/assets/b36c8182-7267-414a-b396-e6e68fa71ff3)