Infrastructure: rationalise font handling (#7918)
#### Brief overview of PR changes/additions
* Revert redesign of main font selection in preferences.
* Restructure usage of fonts to remove the somewhat redundent `(QFont)
mDisplayFont` for widgets that inherently have their own (QFont) which
they use for painting operations.
* Provide a means to track only the details of `QFont`s that we care
about which should be more lightweight than holding a complete copy of a
font (in `TFontDetails`).
* Include the "antialisaing" detall for the main display font (only for
the main console) in the active updates that changing the main font in
the preferences produces.
* Apply changes to the font in a `TConsole` to **all** the
`TCommandLine`s associated with it.

#### Brief overview of PR changes/additions
* The previously revised font settings in the preferences that used a
native font selection was not as useful as it seemed - at least the
Windows one offered controls that we did not need or use which meant
they had not actual effects which would be confusing to the end user.
* Storing a separate copy of `QFont`s did not seem productive especially
as it was easy to modify the intrinsic one and not the copy or
vice-versa.
* Adding a separate class to track the font details we ARE interested in
makes it easier to compare fonts and to tell whether two instances are
the same as far as Mudlet is concerned
* It was not clear that the "anti-aliasing" setting was being correctly
applied/used where it was intended - with this PR it is also updated as
it is changed in the prefernces along with the other two settings: font
"family" and "size".
* The font in command-lines now clearly follow the main console or a
sub- console that they belong to.

#### Other info (issues closed, discussion etc)
Once this is done it is perhaps a bit clearer that the Lua API
`setFont(["windowName", ] "fontName")` sets the font family to use for
the main or (if given a name) a sub-console/user-window - and if it is
the "Main console" that also carries through into:
* the Lua script window in the editor
* the error window in the editor
* the "notepad"
* the map info display in the 2D mapper

the font is also replicated in the corresponding "command-lines" for
that "console".

The Lua API `setFontSize(["windowName", ]integer)` sets the font size
for all the places where the correspond font had been set by the
`setFont(...)` call. Setting the font name and size via the Lua API are
two separate operations and they do act seemingly independently of each
other.

Setting these things via the preferences take effect on the main console
(and related things) immediately (except for the mapper - that only
updates when the preferences dialogue is closed) - the anti-alias
setting also acts on these things in the same manner (which it didn't
before IIRC).

As a side effect of the reversion this will close #7907!

I have a separate commit that gives the option for
sub-console/user-windows to track the main console font family (so that
changing the font in there carries through into those other windows) -
optionally they can *also* track the size of the main font (by a
user/scriptable factor) so that as the main font is made smaller or
larger so do the other consoles - however this can be switched off if
not wanted. I've let that as a separate commit because it would just
enlarge this already meaty PR and it can be safely left to be
reviewed/done separately!

---------

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>