Fix: Remove extra added linebreak for command echoes during trigger processing. (#7939)
#### Brief overview of PR changes/additions
Removed the added linebreak append from command echoes IF the trigger
engine is processing.

#### Motivation for adding to Mudlet
While this behavior was apparently necessary to keep command echoes from
bunching in the past, with the new word-wrapping and append methods it
was leading to an extra linebreak being inserted between command echoes:

![image](https://github.com/user-attachments/assets/876f2acf-8108-446c-870d-0ad6611de944)

Now fixed:

![image](https://github.com/user-attachments/assets/54b685d7-d29d-48a9-898d-a33665190ab8)


#### Other info (issues closed, discussion etc)
Steps to recreate issue:
Make a trigger which sends 2+ commands. Make sure the "Show the text you
sent" preference in "Input" is enabled.

Thanks to @termie for noticing this and helping me chase it down!