Fix: ordering of languages in directory selector (#7954)
#### Brief overview of PR changes/additions
After #7746 which was part of the UI overhaul of the preferences
dialogue the selection of a dictionary to use for spellchecking for a
profile was moved to a `QComboBox` - however that does not provide a
built in method to sort the texts displayed that the previous
`QListWidget` did. This PR arranges entries in order of the (translated)
texts shown for each dictionary language.

#### Motivation for adding to Mudlet
To make the list of dictionary languages be shown in a reasonable
(alphabetical) order.

#### Other info (issues closed, discussion etc)
`QComboBox::setInsertPolicy(...)` was being used but it is irrelevant in
this situation as it only deals with user inserted items added to
"editable" `QComboBoxes`.

I've also taken this opportunity to restore the indication of the source
of the dictionaries used to populate the list (either shipped with
Mudlet {MacOS, Windows and POSIX OSes where Mudlet cannot find system
`Hunspell` dictionary} or provided by the OS {POSIX ones where they
are}.

I've also added a few more dictionaries that Debian have added to the
set that they provide - so that they are named rather than just showing
a `xx_YY` code if seen.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>