fix: Modified DB.lua assertion to allow for default false fields (#7931)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Allow db functions to work with a field that defaults to `false`
#### Motivation for adding to Mudlet
False is a reasonable default for a database field, and others will
undoubtedly encounter this issue in the future.
#### Other info (issues closed, discussion etc)
Discussion occurred on Discord. Testing live on my own system, should be
safe to merge within a few days.