fix: New Triggers, Aliases, and Scripts can now be disabled (#7945)
#### Brief overview of PR changes/additions

This is meant to solve issue #7723. In order to allow triggers, aliases,
and scripts named "New trigger", "New alias", and "New script" to be
disabled, I added new variables and functions in the TTrigger, TAlias,
and TScript classes. These functions allow one to see whether the most
recent save was the "save" that created the trigger/alias/script; if it
is, then while saving, the variable is changed.

#### Motivation for adding to Mudlet

#### Other info (issues closed, discussion etc)

This is my second pull request. Please let me know if there is any way I
can improve it.