Infrastructure: move from squirrel.windows to Clowd.squirrel
#### Brief overview of PR changes/additions
Try a fork of the original Windows installer version to see if that works
when the original is not.

#### Motivation for adding to Mudlet
The squirrel.windows installer builder has not been updated since 2021 and
it has stopped working in Windows 10/11 builds. It fails to extract all
the `.dll`s (Qt and others) needed leaving the "installer" builds (PTBs and
any forthcoming releases) unusable. The replacement proposed here has
been "archived" on GitHub - but it does still represent a later version
compatible with the original.

#### Other info (issues closed, discussion etc)
Because this PR targets the installer building framework it cannot be
tested as a PR from a third-party (my) GH repository. Instead it must be
sent directly to Mudlet's own repository as a new branch, and then the work
flows must be triggered with a manual input of the "Imitate a schedule
build" set to "true" so as to fake a PTB.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>