Add: sandbox loading metadata from packages (#7972)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Sandboxes the Lua environment used when reading data from `config.lua`
in Lua packages.
#### Motivation for adding to Mudlet
Better security.
#### Other info (issues closed, discussion etc)
Other environments (such as the lua formatter) could be applied to this
sandbox as well. It can't be applied to player scripts just yet though
as-is, because it is likely to break existing player-made scripts that
use these functions while posing no security concerns.

---------

Co-authored-by: Vadim Peretokin <vadi2@users.noreply.github.com>