Improve: Allow users to set map background alpha (#7958)
#### Brief overview of PR changes/additions
- Enable users to set an alpha value for the mapper background in:
Preferences > Mapper colors > Background color
- Make transparent/translucent backgrounds actually work (e.g. 0 for a
fully transparent map background).
- Make the 3D mapper use the color set in Preferences > Mapper colors >
Background color (previously it was permanently black no matter what).
- Properly save map background alpha across sessions.

#### Motivation for adding to Mudlet
This unlocks some really neat GUI options as I personally love having a
map overlay on top of the main console.

#### Other info (issues closed, discussion etc)
The 3D mapper is a bit funky since it is based on QOpenGLWidget which
has some limitations:

> [Limitations and Other
Considerations](https://doc.qt.io/qt-6/qopenglwidget.html#limitations-and-other-considerations)

>Putting other widgets underneath and making the QOpenGLWidget
transparent will not lead to the expected results: The widgets
underneath will not be visible. This is because in practice the
QOpenGLWidget is drawn before all other regular, non-OpenGL widgets, and
so see-through type of solutions are not feasible. Other type of
layouts, like having widgets on top of the QOpenGLWidget, will function
as expected.

>When absolutely necessary, this limitation can be overcome by setting
the
[Qt::WA_AlwaysStackOnTop](https://doc.qt.io/qt-6/qt.html#WidgetAttribute-enum)
attribute on the QOpenGLWidget. Be aware however that this breaks
stacking order, for example it will not be possible to have other
widgets on top of the QOpenGLWidget, so it should only be used in
situations where a semi-transparent QOpenGLWidget with other widgets
visible underneath is required.

So I've gone ahead and made it so setting the alpha value of the map
under 255 toggles transparency on AND the 3D map will always stack on
top. The unfortunate side effect is that everything then gets stuck
behind the 3D mapper. Hopefully this is niche enough to not cause
issues. If it is problematic, users are still free to set the alpha back
to 255 and AlwaysStackOnTop will be disabled.

Pardon the light mode!

https://github.com/user-attachments/assets/0c501008-974b-41bd-b1be-ecba288f4c71

To make background transparent as in the video set the alpha to `0` in:
Preferences > Mapper colors > Background color

---------

Co-authored-by: Zooka <136661366+ZookaOnGit@users.noreply.github.com>