Fix: make leak sanitizer happy in i18n (#8328)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Make leak sanitizer happy in i18n code by deleting the objects when
Mudlet is quitting. It's not a leak in a true sense because Mudlet is
shut down anyway, so all memory is freed anyhow, but it does make the
sanitizer happier.
#### Motivation for adding to Mudlet
Cleaner leak sanitizer output.
#### Other info (issues closed, discussion etc)
As a best practice, Qt object should always [have a
parent](https://doc.qt.io/qt-6/object.html). That way, when the parent
is deleted, all the children are too - and thus, there is no memory
leaks.

Co-authored-by: Vadim Peretokin <vadi2@users.noreply.github.com>