Fix: memory leak that happened when running tests (#8364)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Fixed a memory leak that happened when running tests. Qt's deleteLater()
was not being called when the application was shut down, leading to a
false positive to appear. This didn't happen during normal playtime with
Mudlet, so it's not an issue that affected players.
#### Motivation for adding to Mudlet
No memory leaks at all, not even in tests, so our tests can see if new
leaks appear in code or not.
#### Other info (issues closed, discussion etc)